rename network sender actor
