+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/messaginglayer/DeadLetterMonitorActor.scala
@@ -1,7 +1,7 @@
 package edu.uci.ics.amber.engine.architecture.messaginglayer
 
 import akka.actor.{Actor, DeadLetter}
-import edu.uci.ics.amber.engine.architecture.messaginglayer.NetworkSenderActor.{
+import edu.uci.ics.amber.engine.architecture.messaginglayer.NetworkCommunicationActor.{
   MessageBecomesDeadLetter,
   NetworkMessage
 }
